all api endpoints are added

+++ IMDB-CLONE-MOVIESERVICE/src/main/java/org/roybond007/model/dto/MovieInfoResponseBody.java
@@ -24,7 +24,7 @@ public class MovieInfoResponseBody {
 	private long noOfRatings;
 	private double avgRating;
 	private String genres;
-	private double userRating;
+	private long userRating;
 	
 	
 }
